rpm, utils: Move universally useful functions from the rpm repo to utils
Some functionality such as path canonicalization is useful for more types of
repositories besides rpm. As such, move these functions from the rpm repository
to the `utils` module and make them publicy callable.

Additionally, make `calculate_stability` return a f64 directly and include a safety
comment on why the `.iter().min().unwrap()` call won't fail.
